Cleaned up code in wallet encapsulated functions in ChargeCardModal
